backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
openai==1.82.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic[email]==2.10.0
python-multipart==0.0.12
langgraph==0.2.56
langchain-core==0.3.25
faiss-cpu==1.9.0.post1
numpy==1.26.4
sse-starlette==2.1.3
structlog==24.4.0
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1
prometheus-client==0.21.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
